move: a watch folder => in a emulator folder
